# partB/openapi.yaml
openapi: 3.0.3
info:
  title: Library Lending API
  description: |
    Номын сангийн ном зээлэх системийн REST API.
    - Гишүүн нэгэн зэрэг ≤5 ном зээлнэ
    - 14 хоногийн зээлийн хугацаа
    - 1 удаа сунгах боломжтой
    - RFC 7807 алдааны формат
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Problem:
      type: object
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
      required: [type, title, status]

    Book:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        author: { type: string }
        isbn: { type: string }
        genre: { type: string }
        totalCopies: { type: integer }
        availableCopies: { type: integer }
        available: { type: boolean }
        createdAt: { type: string, format: date-time }

    Loan:
      type: object
      properties:
        id: { type: string }
        bookId: { type: string }
        bookTitle: { type: string }
        memberId: { type: string }
        memberName: { type: string }
        borrowedAt: { type: string, format: date-time }
        dueDate: { type: string, format: date-time }
        returnedAt: { type: string, format: date-time, nullable: true }
        status: { type: string, enum: [active, returned] }
        extended: { type: boolean }
        extensionCount: { type: integer }

    Member:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
        role: { type: string, enum: [admin, member] }
        activeLoans: { type: integer }
        createdAt: { type: string, format: date-time }

    Reservation:
      type: object
      properties:
        id: { type: string }
        bookId: { type: string }
        bookTitle: { type: string }
        memberId: { type: string }
        reservedAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        status: { type: string, enum: [active, fulfilled, cancelled] }

    PaginatedBooks:
      type: object
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/Book' } }
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }

security:
  - bearerAuth: []

paths:
  /auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Login and get JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken: { type: string }
                  userId: { type: string }
                  name: { type: string }
                  role: { type: string }
        '401':
          description: Invalid credentials
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }

  /auth/register:
    post:
      tags: [Auth]
      security: []
      summary: Register a new member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                email: { type: string }
                password: { type: string }
      responses:
        '201':
          description: Member created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Member' }
        '409':
          description: Email already in use
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }

  /books:
    get:
      tags: [Books]
      summary: List books with pagination, filtering, sorting
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 10 } }
        - { name: title, in: query, schema: { type: string } }
        - { name: author, in: query, schema: { type: string } }
        - { name: genre, in: query, schema: { type: string } }
        - { name: available, in: query, schema: { type: boolean } }
        - { name: sort, in: query, schema: { type: string, enum: [title, author, createdAt] } }
        - { name: order, in: query, schema: { type: string, enum: [asc, desc] } }
      responses:
        '200':
          description: Paginated book list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginatedBooks' }
    post:
      tags: [Books]
      summary: Add a new book (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, author, isbn]
              properties:
                title: { type: string }
                author: { type: string }
                isbn: { type: string }
                genre: { type: string }
                totalCopies: { type: integer, default: 1 }
      responses:
        '201':
          description: Book created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Book' }
        '400': { description: Validation error }
        '403': { description: Forbidden }
        '409': { description: ISBN conflict }

  /books/{id}:
    get:
      tags: [Books]
      summary: Get a specific book
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Book' }
        '404':
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
    delete:
      tags: [Books]
      summary: Delete a book (admin only)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '204': { description: Deleted }
        '404': { description: Not found }
        '409': { description: Active loans exist }

  /loans:
    get:
      tags: [Loans]
      summary: List loans
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [active, returned] } }
        - { name: page, in: query, schema: { type: integer } }
        - { name: limit, in: query, schema: { type: integer } }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Loan' } }
                  total: { type: integer }
    post:
      tags: [Loans]
      summary: Borrow a book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bookId]
              properties:
                bookId: { type: string }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Loan' }
        '404': { description: Book or member not found }
        '409': { description: Already borrowed or no copies }
        '422': { description: Loan limit exceeded (5 max) }

  /loans/{id}/return:
    patch:
      tags: [Loans]
      summary: Return a borrowed book
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Loan' }
        '404': { description: Loan not found }
        '409': { description: Loan not active }

  /loans/{id}/extend:
    patch:
      tags: [Loans]
      summary: Extend loan by 14 days (once only)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Loan' }
        '422': { description: Already extended once }

  /members:
    get:
      tags: [Members]
      summary: List all members (admin only)
      responses:
        '200': { description: Member list }
        '403': { description: Admin required }

  /reservations:
    get:
      tags: [Reservations]
      summary: List reservations
      responses:
        '200': { description: Reservation list }
    post:
      tags: [Reservations]
      summary: Reserve a book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bookId]
              properties:
                bookId: { type: string }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Reservation' }
        '409': { description: Book available or already reserved }

  /reservations/{id}:
    delete:
      tags: [Reservations]
      summary: Cancel a reservation
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '204': { description: Cancelled }
        '404': { description: Not found }